+++ src/commands/meme.js
@@ -1,5 +1,6 @@
 const { SlashCommandBuilder } = require('@discordjs/builders');
-const { MessageEmbed } = require('discord.js');
+const { MessageEmbed, MessageAttachment } = require('discord.js');
+const { defaultColor } = require('../../config');
 const fetch = require('node-fetch');
 
 module.exports = {
@@ -8,8 +9,11 @@ module.exports = {
 		.setDescription('Send a random meme from r/meme'),
 	async execute(interaction) {
 		const data = await fetch('http://meme-api.herokuapp.com/gimme').then(res => res.json());
+		const image = new MessageAttachment('./assets/funny.png', 'funny.png');
 
 		const embed = new MessageEmbed()
+			.setColor(defaultColor)
+			.setThumbnail('attachment://funny.png')
 			.setAuthor(data.author, 'https://www.redditstatic.com/avatars/defaults/v2/avatar_default_1.png', 'https://reddit.com/u/' + data.author)
 			.setTitle(data.title)
 			.setURL(data.postLink)
@@ -17,6 +21,6 @@ module.exports = {
 			.setFooter(`r/${data.subreddit} • ${data.ups} 👍`)
 			.setTimestamp();
 
-		interaction.reply({ embeds: [embed] });
+		interaction.reply({ embeds: [embed], files: [image] });
 	},
 };