+++ src/commands/reddit.js
@@ -1,5 +1,5 @@
 const { SlashCommandBuilder } = require('@discordjs/builders');
-const { MessageEmbed } = require('discord.js');
+const { MessageEmbed, MessageAttachment } = require('discord.js');
 const fetch = require('node-fetch');
 
 module.exports = {
@@ -10,6 +10,7 @@ module.exports = {
 	async execute(interaction) {
 		const subreddit = interaction.options.getString('subreddit');
 		const data = await fetch('http://meme-api.herokuapp.com/gimme/' + subreddit).then(res => res.json());
+		const image = new MessageAttachment('./assets/reddit.png', 'reddit.png');
 
 		if (data.message == `r/${subreddit.toLowerCase()} has no Posts with Images`) {
 			return interaction.reply({ content: data.message, ephemeral: true });
@@ -17,6 +18,7 @@ module.exports = {
 
 		const embed = new MessageEmbed()
 			.setColor('#FF4300')
+			.setThumbnail('attachment://reddit.png')
 			.setAuthor(data.author, 'https://www.redditstatic.com/avatars/defaults/v2/avatar_default_1.png', 'https://reddit.com/u/' + data.author)
 			.setTitle(data.title)
 			.setURL(data.postLink)
@@ -24,6 +26,6 @@ module.exports = {
 			.setFooter(`r/${data.subreddit} • ${data.ups} 👍`)
 			.setTimestamp();
 
-		interaction.reply({ embeds: [embed] });
+		interaction.reply({ embeds: [embed], files: [image] });
 	},
 };